requirements.txt
astropy==6.0.0
Cartopy==0.22.0
dash==2.18.2
dash_core_components==2.0.0
dash_gif_component==1.1.0
dash_html_components==2.0.0
geopack==1.0.10
gpytorch==1.11
imageio==2.37.0
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
Pillow==11.1.0
plotly==6.0.0
pyspedas==1.5.6
pytplot_mpl_temp==2.2.6
Requests==2.32.3
rich==13.9.4
scikit_learn==1.3.0
scipy==1.15.1
streamlit==1.42.0
sunpy==5.1.1
torch==2.3.0+cu121
tqdm==4.65.0
spacepy==0.7.0
bezpy==0.1.1
aacgmv2==2.7.0